Initial commit. Some HTML started.
Create basic structure of a website that consists of 3 relatively simple sections, each one similar in structure. Also include navigation that utilizes a hamburger menu. The focus of this project will be on experimenting with designs that are very minimalistic, perhaps making use of strong animations and transitions to accompany the lack of illustrations and text.