Begin adding some transitions and scroll events
Using MagicScroll and GSAP, start adding some transitions for each section of the page. The transitions so far include the image and paragraphs being hidden at first, and then being completely unobscured. Next, using MagicScroll, I make these transitions only occur when the user scrolls down the page (or on page load for the first section). Finally, an overall animation is added to each section, causing it to shrink and fade out as the user scrolls away from it.